+++ CHANGELOG.md
@@ -29,6 +29,7 @@ All notable changes to this project will be documented in this file.
 * Fix `DelegateProxy` call to `layoutIfNeeded` for an object without a window. #2076
 * Add `ControlEvent` wrappers to `UIApplication` Notifications. #2116
 * `SharingScheduler.mock(scheduler:action:)` can use throwing function for `action`. #2150
+* Add `interlocked: Bool = false` parameter to `VirtualTimeScheduler.sleep`.
 
 ## [5.1.1](https://github.com/ReactiveX/RxSwift/releases/tag/5.1.1)